// types/records.ts
export type TreeRecord = {
  _id: string;
  "formhub/uuid": string;
  i_suanid: string;
  "a_data/a_username": string;
  "a_data/a_conglomerado": string;
  "a_data/a_submission_date": string;
  "a_data/a_subparcela": string;
  "a_data/a_no_ia_individuo": string;
  "a_data/a_category": string;
  "a_data/a_identificacion_arbol": string;
  "a_data/a_geopoint_map": string;
  "b_data/b_nombre_comun": string;
  "b_data/b_especie": string;
  "b_data/b_familia": string;
  "b_data/b_foto_arbol": string;
  "b_data/b_foto_tronco": string;
  "b_data/b_foto_hojas": string;
  "c_data/c_arbol_vivo": string;
  "c_data/c_estado_arbol": string;
  "c_data/c_foto_arbol_afectada": string;
  "d_data/d_tallo_tipo": string;
  "d_data/d_equipo_1": string;
  "d_data/d_diametro_h_fuste_1": string;
  "d_data/d_observacion": string;
  "d_data/d_distancia": string;
  "d_data/d_equipo_2": string;
  "d_data/d_ht_arbol_fuste_1": string;
  _submission_time: string;
  _submitted_by: string;
  _geolocation: string;
  [key: string]: string | null;
}; 

export type TransformedRecord = {
  id: string;
  suanId: string;
  username: string;
  conglomerado: string;
  fecha: string;
  subparcela: string;
  noIndividuo: string;
  categoria: string;
  identificacionArbol: string;
  geopoint: string;
  nombreComun: string;
  especie: string;
  familia: string;
  fotoArbol: string;
  fotoTronco: string;
  fotoHojas: string;
  arbolVivo: string;
  estadoArbol: string;
  fotoArbolAfectada: string;
  talloTipo: string;
  equipo1: string;
  diametroFuste: string;
  observacion: string;
  distancia: string;
  equipo2: string;
  alturaArbol: string;
  submissionTime: string;
  submittedBy: string;
  geolocation: string;
};